feat: initiate data wrangle project and gather data from Twitter api
Initiated the Udacity data wrangle and analyse project, where using
Python and its libraries, I will gather data from a variety of sources
and in a variety of formats, assess its quality and tidiness,
then clean it.

I will document my wrangling efforts in a Jupyter Notebook, plus
showcase them through analyses and visualizations
using Python (and its libraries).

The dataset that I will be wrangling (and analyzing and visualizing) is
the tweet archive of Twitter user @dog_rates, also known as WeRateDogs.
WeRateDogs is a Twitter account that rates people's dogs with a
humorous comment about the dog.